+++ packages/amplify-graphql-http-transformer/src/index.ts
@@ -1,0 +1,1 @@
+export { HttpTransformer } from './graphql-http-transformer';