+++ packages/amplify-provider-awscloudformation/src/graphql-transformer/transform-graphql-schema.ts
@@ -11,7 +11,8 @@ import {
   TransformerProjectConfig,
 } from '@aws-amplify/graphql-transformer-core';
 import { ModelTransformer } from '@aws-amplify/graphql-model-transformer';
-
+import { FunctionTransformer } from '@aws-amplify/graphql-function-transformer';
+import { HttpTransformer } from '@aws-amplify/graphql-http-transformer';
 import { ProviderName as providerName } from '../constants';
 import { hashDirectory } from '../upload-appsync-files';
 import { writeDeploymentToDisk } from './utils';
@@ -46,6 +47,8 @@ function getTransformerFactory(context, resourceDir) {
   return async () => {
     const transformerList: TransformerPluginProvider[] = [
       new ModelTransformer(),
+      new FunctionTransformer(),
+      new HttpTransformer(),
       // TODO: initialize transformer plugins
     ];