+++ packages/amplify-frontend-ios/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.19.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.19.4...amplify-frontend-ios@2.19.6) (2021-05-03)
+
+
+
+## 4.50.1 (2021-05-03)
+
+
+### Bug Fixes
+
+* render the right auth mode for admin queries in config files ([#7118](https://github.com/aws-amplify/amplify-cli/issues/7118)) ([8f8deba](https://github.com/aws-amplify/amplify-cli/commit/8f8deba9171a30433aa6820af1f32f5498db5028)), closes [#6983](https://github.com/aws-amplify/amplify-cli/issues/6983)
+
+
+
+
+
+## [2.19.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.19.4...amplify-frontend-ios@2.19.5) (2021-05-03)
+
+
+### Bug Fixes
+
+* render the right auth mode for admin queries in config files ([#7118](https://github.com/aws-amplify/amplify-cli/issues/7118)) ([8f8deba](https://github.com/aws-amplify/amplify-cli/commit/8f8deba9171a30433aa6820af1f32f5498db5028)), closes [#6983](https://github.com/aws-amplify/amplify-cli/issues/6983)
+
+
+
+
+
+## [2.19.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.19.3...amplify-frontend-ios@2.19.4) (2021-04-27)
+
+
+### Bug Fixes
+
+* **amplify-frontend-ios:** update xcode project on pull ([#7152](https://github.com/aws-amplify/amplify-cli/issues/7152)) ([9d5cb21](https://github.com/aws-amplify/amplify-cli/commit/9d5cb21b08ebc6a09c5c005e8130f4ba7db757ea))
+
+
+
+
+
 ## [2.19.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.19.2...amplify-frontend-ios@2.19.3) (2021-04-19)
 
 **Note:** Version bump only for package amplify-frontend-ios