+++ packages/amplify-graphiql-explorer/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@1.4.3...amplify-graphiql-explorer@1.6.0) (2021-05-03)
+
+
+
+## 4.50.1 (2021-05-03)
+
+
+### Features
+
+* add mock support for custom fields in cognito jwt ([#7172](https://github.com/aws-amplify/amplify-cli/issues/7172)) ([e4210cc](https://github.com/aws-amplify/amplify-cli/commit/e4210cc25397a92ae0d6ce5a9d5f4ecc5da8ae6f)), closes [#6106](https://github.com/aws-amplify/amplify-cli/issues/6106)
+
+
+
+
+
+# [1.5.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@1.4.3...amplify-graphiql-explorer@1.5.0) (2021-05-03)
+
+
+### Features
+
+* add mock support for custom fields in cognito jwt ([#7172](https://github.com/aws-amplify/amplify-cli/issues/7172)) ([e4210cc](https://github.com/aws-amplify/amplify-cli/commit/e4210cc25397a92ae0d6ce5a9d5f4ecc5da8ae6f)), closes [#6106](https://github.com/aws-amplify/amplify-cli/issues/6106)
+
+
+
+
+
 ## [1.4.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@1.4.2...amplify-graphiql-explorer@1.4.3) (2021-03-05)
 
 **Note:** Version bump only for package amplify-graphiql-explorer